Framing function is finally working as expected.
The gm node.js module seems to do some mangling of the imagemagick
-border option. I couldn't get anything other than a giant border (?!).
Specifying it manually seems to work.

I've also got a handle on the label option, reading the gm docs
(which are impossible to find). There doesn't appear to be any
fonts bundled in, which is fine as I want to use my own anyways.

Got to fix the centering. Maybe the border needs to be more exciting too