Make the database transaction idempotent
The google cloud functions are guaranteed to be called "at least once".
That is to say, they may be called twice for a single event trigger. To
ensure that I don't end up with multiple database entries for the same
file, key them on the filename (minus the extension), as these need to
be unique anyway.